refactor(architecture-diagram): extract layout and checks modules
Move deterministic placement, routing, geometry validation, and deployment profile checks behind owner modules.

Stack-Id: architecture-diagram-m1-3d8ce7
Stack-Position: 2/4
